logInfo

Log information with time stamp accounting for message type. Message type will point toward the most appropriate cli function display.

Usage

logInfo(msg, type = "info")

Arguments

msg

Character values of message to log that leverages cli formatting.

type

Name of the message type to toward best cli display:

"info": uses cli::cli_alert_info()

"success": uses cli::cli_alert_success()

"h1": uses cli::cli_h1()

"h2": uses cli::cli_h2()

"h3": uses cli::cli_h3()

"text": uses cli::cli_text()

"alert": uses cli::cli_alert()

"li": uses cli::cli_li()

"ol": uses cli::cli_ol()

"progress_step": uses cli::cli_progress_step()

Examples

# Log information
logInfo(cliFormat("This is an {.strong info} message"))
#> ℹ Info [25/07/2025 - 09:29:48]:  This is an info message
#> 

# Log a title
logInfo(cliFormat("Task: {.strong tic toc test}"), type = "h1")
#> 
#> ── Task: tic toc test ──────────────────────────────────────────────────────────

# Log success
t0 <- tic()
Sys.sleep(3)
logInfo(cliFormat("Task: {.strong tic toc test} completed [{toc(t0, \"s\")}]"), type = "success")
#> ✔ Info [25/07/2025 - 09:29:51]:  Task: tic toc test completed [3.0 s]
#>